Chore(styling) Moved styled components to new file and created a Home folder.
