+++ client/src/App.js
@@ -2,7 +2,7 @@ import { useState } from "react";
 import { useNavigate } from "react-router-dom";
 //components
 
-import Home from "./components/Home";
+import Home from "./components/Home/Home";
 import Login from "./components/Login/Login";
 //routes
 import { Routes, Route } from "react-router-dom";
