+++ client/src/components/Home/Jumbotron.js
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import BackgroundImage from "../assets/BackgroundImage.jpeg";
+import BackgroundImage from "../../assets/BackgroundImage.jpeg";
 
 const JumbotronContainer = styled.div`
   height: 400px;
